Refactor show module into `show-details` & `discover`
