+++ shared/domain/discover/api/build.gradle.kts
