+++ shared/domain/discover/api/src/androidMain/AndroidManifest.xml
