+++ shared/domain/discover/implementation/build.gradle.kts
@@ -1,0 +1,23 @@
+plugins {
+    `kmm-domain-plugin`
+}
+
+dependencies {
+    commonMainImplementation(project(":shared:remote"))
+    commonMainImplementation(project(":shared:domain:discover:api"))
+    commonMainImplementation(project(":shared:domain:show-common:api"))
+
+    commonMainImplementation(libs.koin.core)
+    commonMainImplementation(libs.kotlin.datetime)
+    commonMainImplementation(libs.multiplatform.paging.core)
+    commonMainImplementation(libs.squareup.sqldelight.extensions)
+
+    testImplementation(libs.testing.mockk.core)
+
+    commonTestImplementation(kotlin("test"))
+    commonTestImplementation(project(":shared:core-test"))
+    commonTestImplementation(libs.testing.turbine)
+    commonTestImplementation(libs.testing.kotest.assertions)
+    commonTestImplementation(libs.testing.coroutines.test)
+    commonTestImplementation(libs.testing.mockk.common)
+}
