+++ shared/domain/discover/implementation/src/androidMain/AndroidManifest.xml
