+++ shared/domain/discover/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/discover/implementation/DiscoverCategoryCacheImpl.kt
@@ -1,0 +1,29 @@
+package com.thomaskioko.tvmaniac.discover.implementation
+
+import com.squareup.sqldelight.runtime.coroutines.asFlow
+import com.squareup.sqldelight.runtime.coroutines.mapToList
+import com.thomaskioko.tvmaniac.datasource.cache.SelectShows
+import com.thomaskioko.tvmaniac.datasource.cache.Show_category
+import com.thomaskioko.tvmaniac.datasource.cache.TvManiacDatabase
+import com.thomaskioko.tvmaniac.discover.api.cache.DiscoverCategoryCache
+import kotlinx.coroutines.flow.Flow
+
+class DiscoverCategoryCacheImpl(
+    private val database: TvManiacDatabase
+) : DiscoverCategoryCache {
+
+    private val showCategoryQuery get() = database.showCategoryQueries
+
+    override fun insert(category: Show_category) {
+        showCategoryQuery.insertOrReplace(
+            show_id = category.show_id,
+            category_id = category.category_id
+        )
+    }
+
+    override fun observeShowsByCategoryID(id: Int): Flow<List<SelectShows>> {
+        return showCategoryQuery.selectShows(category_id = id.toLong())
+            .asFlow()
+            .mapToList()
+    }
+}
