+++ shared/domain/show-details/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/details/api/cache/ShowCategoryCache.kt
@@ -1,0 +1,8 @@
+package com.thomaskioko.tvmaniac.details.api.cache
+
+import com.thomaskioko.tvmaniac.datasource.cache.SelectShows
+
+interface ShowCategoryCache {
+
+    fun getShowsByCategoryID(id: Int): List<SelectShows>
+}
