+++ shared/domain/show-details/implementation/build.gradle.kts
@@ -1,0 +1,25 @@
+plugins {
+    `kmm-domain-plugin`
+}
+
+dependencies {
+    commonMainImplementation(projects.shared.remote)
+    commonMainImplementation(projects.shared.domain.showDetails.api)
+    commonMainImplementation(projects.shared.domain.episodes.api)
+    commonMainImplementation(projects.shared.domain.lastAirEpisodes.api)
+    commonMainImplementation(projects.shared.domain.showCommon.api)
+
+    commonMainImplementation(libs.koin.core)
+    commonMainImplementation(libs.kotlin.datetime)
+    commonMainImplementation(libs.multiplatform.paging.core)
+    commonMainImplementation(libs.squareup.sqldelight.extensions)
+
+    testImplementation(libs.testing.mockk.core)
+
+    commonTestImplementation(kotlin("test"))
+    commonTestImplementation(projects.shared.coreTest)
+    commonTestImplementation(libs.testing.turbine)
+    commonTestImplementation(libs.testing.kotest.assertions)
+    commonTestImplementation(libs.testing.coroutines.test)
+    commonTestImplementation(libs.testing.mockk.common)
+}
