+++ shared/domain/show-details/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/details/implementation/cache/ShowCategoryCacheImpl.kt
@@ -1,0 +1,18 @@
+package com.thomaskioko.tvmaniac.details.implementation.cache
+
+import com.thomaskioko.tvmaniac.datasource.cache.SelectShows
+import com.thomaskioko.tvmaniac.datasource.cache.TvManiacDatabase
+import com.thomaskioko.tvmaniac.details.api.cache.ShowCategoryCache
+
+class ShowCategoryCacheImpl(
+    private val database: TvManiacDatabase
+) : ShowCategoryCache {
+
+    private val showCategoryQuery get() = database.showCategoryQueries
+
+    override fun getShowsByCategoryID(id: Int): List<SelectShows> {
+        return showCategoryQuery.selectShows(
+            category_id = id.toLong()
+        ).executeAsList()
+    }
+}
